+++ ui/app/utils/clickhouse/evaluations.ts
@@ -4,8 +4,6 @@ import {
   jsonInferenceOutputSchema,
   resolvedInputSchema,
 } from "./common";
-import { inputSchema } from "./common";
-import { resolveInput } from "../resolve.server";
 
 export const EvaluationRunInfoSchema = z.object({
   eval_run_id: z.string(),
@@ -93,76 +91,6 @@ export const ParsedEvaluationResultWithVariantSchema = z.union([
 export type ParsedEvaluationResultWithVariant = z.infer<
   typeof ParsedEvaluationResultWithVariantSchema
 >;
-
-export async function parseEvaluationResult(
-  result: EvaluationResult,
-): Promise<ParsedEvaluationResult> {
-  try {
-    // Parse the input field
-    const parsedInput = inputSchema.parse(JSON.parse(result.input));
-    const resolvedInput = await resolveInput(parsedInput);
-
-    // Parse the outputs
-    const generatedOutput = JSON.parse(result.generated_output);
-    const referenceOutput = JSON.parse(result.reference_output);
-
-    // Determine if this is a chat result by checking if generated_output is an array
-    if (Array.isArray(generatedOutput)) {
-      // This is likely a chat evaluation result
-      return ChatEvaluationResultSchema.parse({
-        ...result,
-        input: resolvedInput,
-        generated_output: generatedOutput,
-        reference_output: referenceOutput,
-      });
-    } else {
-      // This is likely a JSON evaluation result
-      return JsonEvaluationResultSchema.parse({
-        ...result,
-        input: resolvedInput,
-        generated_output: generatedOutput,
-        reference_output: referenceOutput,
-      });
-    }
-  } catch (error) {
-    console.warn(
-      "Failed to parse evaluation result using structure-based detection:",
-      error,
-    );
-    // If structure-based detection fails, try the original parsing as fallback
-    return ParsedEvaluationResultSchema.parse(result);
-  }
-}
-
-export async function parseEvaluationResultWithVariant(
-  result: EvaluationResultWithVariant,
-): Promise<ParsedEvaluationResultWithVariant> {
-  try {
-    // Parse using the same logic as parseEvaluationResult
-    const parsedResult = await parseEvaluationResult(result);
-
-    // Add the variant_name to the parsed result
-    const parsedResultWithVariant = {
-      ...parsedResult,
-      variant_name: result.variant_name,
-    };
-    return ParsedEvaluationResultWithVariantSchema.parse(
-      parsedResultWithVariant,
-    );
-  } catch (error) {
-    console.warn(
-      "Failed to parse evaluation result with variant using structure-based detection:",
-      error,
-    );
-    // Fallback to direct parsing if needed
-    return ParsedEvaluationResultWithVariantSchema.parse({
-      ...result,
-      input: result.input,
-      generated_output: result.generated_output,
-      reference_output: result.reference_output,
-    });
-  }
-}
 
 export const EvaluationStatisticsSchema = z.object({
   eval_run_id: z.string(),
